Encrypted File Transfer/requirements.txt
streamlit
pycryptodome
cryptography
